Deploying the image to Digital Ocean
In this commit, we add two scripts using the `s3cmd` and `doctl` tools.
They use the credentials introduced in the previous commit: `s3cmd`
uploads to S3 the image built in the commit before, and `doctl` is used
to import the image into Digital Ocean then spin a new virtual machine.